+++ xyz.struthers.rhul.ham/src/main/java/xyz/struthers/rhul/ham/agent/AustralianGovernment.java
@@ -10,7 +10,7 @@ import java.util.Map;
 import xyz.struthers.rhul.ham.process.NodePayment;
 
 /**
- * Each instance of this class stores 19 doubles, so will consume approximately
+ * Each instance of this class stores 19 floats, so will consume approximately
  * 152 bytes of RAM. There is only one instance of this class in the model.
  * 
  * @author Adam Struthers
@@ -22,32 +22,32 @@ public final class AustralianGovernment extends Agent {
 
 	// agernt relationships
 	protected int paymentClearingIndex;
-	
+
 	// P&L
-	private double pnlTaxIncome;
-	private double pnlSaleOfGoodsAndServices;
-	private double pnlInterestIncome;
-	private double pnlOtherIncome;
+	private float pnlTaxIncome;
+	private float pnlSaleOfGoodsAndServices;
+	private float pnlInterestIncome;
+	private float pnlOtherIncome;
 
-	private double pnlPersonnelExpenses;
-	private double pnlInterestExpense;
-	private double pnlDepreciationAmortisation;
-	private double pnlOtherExpenses;
+	private float pnlPersonnelExpenses;
+	private float pnlInterestExpense;
+	private float pnlDepreciationAmortisation;
+	private float pnlOtherExpenses;
 
-	private double pnlNetAcquisitionOfNonFinancialAssets;
+	private float pnlNetAcquisitionOfNonFinancialAssets;
 
 	// Bal Sht
-	private double bsCash;
-	private double bsInvestmentsLoansPlacements;
-	private double bsEquityAssets;
-	private double bsOtherFinancialAssets;
-	private double bsLandAndFixedAssets;
-	private double bsOtherNonFinancialAssets;
+	private float bsCash;
+	private float bsInvestmentsLoansPlacements;
+	private float bsEquityAssets;
+	private float bsOtherFinancialAssets;
+	private float bsLandAndFixedAssets;
+	private float bsOtherNonFinancialAssets;
 
-	private double bsCurrencyOnIssue;
-	private double bsDepositsHeld;
-	private double bsBorrowings;
-	private double bsOtherLiabilities;
+	private float bsCurrencyOnIssue;
+	private float bsDepositsHeld;
+	private float bsBorrowings;
+	private float bsOtherLiabilities;
 
 	/**
 	 * Default constructor
@@ -63,7 +63,7 @@ public final class AustralianGovernment extends Agent {
 	 * @param balSht
 	 * @param profitLoss
 	 */
-	public AustralianGovernment(Map<String, Double> balSht, Map<String, Double> profitLoss) {
+	public AustralianGovernment(Map<String, Float> balSht, Map<String, Float> profitLoss) {
 		super();
 		this.init();
 
@@ -115,34 +115,34 @@ public final class AustralianGovernment extends Agent {
 		return null;
 	}
 
-	public Map<String, Double> getFinancialStatements() {
+	public Map<String, Float> getFinancialStatements() {
 		// initialised to the number of fields in this class
-		Map<String, Double> result = new HashMap<String, Double>(19);
+		Map<String, Float> result = new HashMap<String, Float>(19);
 
 		// P&L
 		result.put("Taxation revenue", this.pnlTaxIncome);
 		result.put("Sale of goods and services", this.pnlSaleOfGoodsAndServices);
-		result.put("Interest income", this.pnlInterestIncome );
+		result.put("Interest income", this.pnlInterestIncome);
 		result.put("Other income", this.pnlOtherIncome);
 
-		result.put("Employee expenses", this.pnlPersonnelExpenses );
-		result.put("Interest expense",this.pnlInterestExpense);
-		result.put("Depreciation",this.pnlDepreciationAmortisation );
+		result.put("Employee expenses", this.pnlPersonnelExpenses);
+		result.put("Interest expense", this.pnlInterestExpense);
+		result.put("Depreciation", this.pnlDepreciationAmortisation);
 		result.put("Other expenses", this.pnlOtherExpenses);
 
 		result.put("Total net acquisition of non-financial assets", this.pnlNetAcquisitionOfNonFinancialAssets);
 
 		// Bal Sht
-		result.put("Cash and deposits",this.bsCash );
-		result.put("Investments, loans and placements",this.bsInvestmentsLoansPlacements );
-		result.put("Equity",this.bsEquityAssets );
+		result.put("Cash and deposits", this.bsCash);
+		result.put("Investments, loans and placements", this.bsInvestmentsLoansPlacements);
+		result.put("Equity", this.bsEquityAssets);
 		result.put("Other financial assets", this.bsOtherFinancialAssets);
-		result.put("Land and fixed assets",this.bsLandAndFixedAssets );
-		result.put("Other non-financial assets",this.bsOtherNonFinancialAssets );
+		result.put("Land and fixed assets", this.bsLandAndFixedAssets);
+		result.put("Other non-financial assets", this.bsOtherNonFinancialAssets);
 
-		result.put("Currency on issue",this.bsCurrencyOnIssue );
-		result.put("Deposits held",this.bsDepositsHeld );
-		result.put("Borrowing",this.bsBorrowings );
+		result.put("Currency on issue", this.bsCurrencyOnIssue);
+		result.put("Deposits held", this.bsDepositsHeld);
+		result.put("Borrowing", this.bsBorrowings);
 		result.put("Other liabilities", this.bsOtherLiabilities);
 
 		return result;
@@ -154,30 +154,30 @@ public final class AustralianGovernment extends Agent {
 		super.name = "Australian Government";
 
 		// P&L
-		this.pnlTaxIncome = 0d;
-		this.pnlSaleOfGoodsAndServices = 0d;
-		this.pnlInterestIncome = 0d;
-		this.pnlOtherIncome = 0d;
+		this.pnlTaxIncome = 0f;
+		this.pnlSaleOfGoodsAndServices = 0f;
+		this.pnlInterestIncome = 0f;
+		this.pnlOtherIncome = 0f;
 
-		this.pnlPersonnelExpenses = 0d;
-		this.pnlInterestExpense = 0d;
-		this.pnlDepreciationAmortisation = 0d;
-		this.pnlOtherExpenses = 0d;
+		this.pnlPersonnelExpenses = 0f;
+		this.pnlInterestExpense = 0f;
+		this.pnlDepreciationAmortisation = 0f;
+		this.pnlOtherExpenses = 0f;
 
-		this.pnlNetAcquisitionOfNonFinancialAssets = 0d;
+		this.pnlNetAcquisitionOfNonFinancialAssets = 0f;
 
 		// Bal Sht
-		this.bsCash = 0d;
-		this.bsInvestmentsLoansPlacements = 0d;
-		this.bsEquityAssets = 0d;
-		this.bsOtherFinancialAssets = 0d;
-		this.bsLandAndFixedAssets = 0d;
-		this.bsOtherNonFinancialAssets = 0d;
+		this.bsCash = 0f;
+		this.bsInvestmentsLoansPlacements = 0f;
+		this.bsEquityAssets = 0f;
+		this.bsOtherFinancialAssets = 0f;
+		this.bsLandAndFixedAssets = 0f;
+		this.bsOtherNonFinancialAssets = 0f;
 
-		this.bsCurrencyOnIssue = 0d;
-		this.bsDepositsHeld = 0d;
-		this.bsBorrowings = 0d;
-		this.bsOtherLiabilities = 0d;
+		this.bsCurrencyOnIssue = 0f;
+		this.bsDepositsHeld = 0f;
+		this.bsBorrowings = 0f;
+		this.bsOtherLiabilities = 0f;
 	}
 
 }
